refactor(op-devstack): route single-chain interop presets through opcm.upgrade
Makes NewSingleChainInteropSupernodeProofs / NewSingleChainInteropIsthmusSuper
install super-root dispute games via OPCMv2.upgrade (with
SUPER_ROOT_GAMES_MIGRATION enabled at deploy) instead of opcmMigrator.migrate.
migrate stays in use for multi-chain runtimes where it actually does its job —
moving multiple chains onto a shared DisputeGameFactory.

Consolidates the upgrade runtime/helper into multichain_proofs.go next to the
existing attachSupernodeSuperProofs, drops the now-redundant
NewSingleChainInteropSupernodeProofsViaUpgrade preset, and points the new
TestSuperRootGamesInstalledViaOPCMUpgrade acceptance test at the canonical
single-chain preset so existing single-chain interop tests pick the change up
without any callsite changes.
