+++ op-devstack/sysgo/multichain_proofs.go
@@ -120,6 +120,30 @@ func attachSupernodeSuperProofs(t devtest.T, runtime *MultiChainRuntime, cfg Pre
 	return runtime
 }
 
+// attachSupernodeSuperProofsViaUpgrade is the opcm.upgrade counterpart to
+// attachSupernodeSuperProofs. It installs the super-root dispute games through
+// the standard OPCMv2 upgrade entrypoint, then wires the interop challenger and
+// super proposer exactly like the migration-based helper does.
+func attachSupernodeSuperProofsViaUpgrade(t devtest.T, runtime *MultiChainRuntime, cfg PresetConfig) *MultiChainRuntime {
+	chains := orderedRuntimeChains(runtime)
+	t.Require().NotEmpty(chains, "supernode superproofs runtime must contain at least one chain")
+	t.Require().NotNil(runtime.Supernode, "supernode superproofs runtime must provide a supernode")
+
+	proofChain := chains[0]
+	cls := make([]L2CLNode, 0, len(chains))
+	for _, chain := range chains {
+		t.Require().NotNil(chain, "runtime chain entry must not be nil")
+		cls = append(cls, chain.CL)
+	}
+
+	superrootTime := awaitSuperrootTime(t, cls...)
+	superRoot := getSupernodeSuperRoot(t, runtime.Supernode, superrootTime)
+	upgradeToSuperRoots(t, runtime.Keys, runtime.Migration, runtime.L1Network.ChainID(), runtime.L1EL, superRoot, superrootTime, proofChain.Network.ChainID())
+
+	attachSuperChallengerAndProposer(t, runtime, cfg, gameTypes.SuperCannonGameType)
+	return runtime
+}
+
 // attachSuperChallengerAndProposer wires an interop challenger and a super
 // proposer for proposerGameType into a supernode-backed runtime.
 func attachSuperChallengerAndProposer(
@@ -187,13 +211,28 @@ func NewTwoL2SupernodeProofsRuntimeWithConfig(t devtest.T, interopAtGenesis bool
 	return attachSupernodeSuperProofs(t, runtime, cfg)
 }
 
+// NewSingleChainSupernodeProofsRuntimeWithConfig deploys a single-chain interop
+// runtime and switches it over to super-root dispute games via opcm.upgrade.
+//
+// The OPCM is deployed with SUPER_ROOT_GAMES_MIGRATION enabled, which puts
+// SuperPermissionedCannon in the permissioned slot at genesis (the standard
+// output→super switch). Once a real super root is available, opcm.upgrade adds
+// SUPER_CANNON / SUPER_CANNON_KONA, overrides the starting anchor root, and
+// sets the respected game type to SUPER_CANNON.
+//
+// Unlike opcmMigrator.migrate — which is about moving multiple chains onto a
+// shared DisputeGameFactory and stays in use for the multi-chain runtime —
+// opcm.upgrade is the standard per-chain upgrade entrypoint, which is what
+// single-chain interop tests need.
 func NewSingleChainSupernodeProofsRuntimeWithConfig(t devtest.T, interopAtGenesis bool, cfg PresetConfig) *MultiChainRuntime {
-	cfg = withSuperProofsDeployerFeature(cfg)
+	cfg = withSuperRootGamesAtGenesisDeployerFeatures(cfg)
 	runtime := newSingleChainSupernodeRuntimeWithConfig(t, interopAtGenesis, cfg)
-	chain := runtime.Chains["l2a"]
-	chain.Proposer = startMinimalProposer(t, runtime.Keys, chain.Network, runtime.L1EL, chain.CL, cfg.ProposerOptions...)
+	// No minimal output-root proposer: with SUPER_ROOT_GAMES_MIGRATION enabled
+	// the chain has no PermissionedCannon game at genesis, so a legacy proposer
+	// would just loop on NoImplementation reverts. The super proposer is
+	// attached by attachSupernodeSuperProofsViaUpgrade.
 	attachTestSequencerToRuntime(t, runtime, "dev")
-	return attachSupernodeSuperProofs(t, runtime, cfg)
+	return attachSupernodeSuperProofsViaUpgrade(t, runtime, cfg)
 }
 
 func startSuperProposer(
